Add capability for MATLAB structures and cells in deserialise()
This edit adds deserialisation capability of structures and cells to the matlab_serialise.py file. This is required to store certain combined experiment types, although might be mitigated via proper use of metadata.

Note that this required two other changes - remove typehint for return (now dict allowed for stuct) and add optional argument to return the length of the data extracted (for iterating over the dataset).